feat(prim_geo): 补齐目录几何隐含范围推导与 SCTO 切点，收敛层级查询数据源
gen-model 侧的 cate_single / cata_model 需要 derive_implied_extent_from_axis_map 与 SCTorus::tangent_points；同时移除已由 pe_owner 图查询取代的 TreeIndex provider 与不再使用的 mbd 模块，连带去掉 indextree 依赖。

+++ src/tool/db_tool.rs
@@ -43,9 +43,13 @@ pub fn convert_to_hash(bytes: &[u8]) -> i32 {
     i32::from_be_bytes(bytes.try_into().unwrap())
 }
 
+/// `PDMS_Hash::IsUDA`：`hash > 0x171FAD39`（**无符号**比较）。
+///
+/// 不可写成有符号 `hash > 0x171FAD39`：`0xFFF?xxxx` 派生 UDA 的高位为 1，
+/// 作为 `i32` 为负，会被误判为非 UDA（sam7200 `23584/128` 曾因此 `uda_atts=0`）。
 #[inline]
 pub fn is_uda(hash: i32) -> bool {
-    hash > 0x171FAD39
+    (hash as u32) > 0x171FAD39
 }
 
 #[inline]
